PYIC-5875: Reorganise types
This moves all the interface and type definitions into the types dir. It
also groups them in files bases around the client interactions.
