+++ .gitignore
@@ -57,3 +57,4 @@ secrets.gradle
 api-tests/cucumber-report.html
 api-tests/.npmrc
 api-tests/.env
+api-tests/src/**/*.js
