--- api-tests/src/interfaces/auth-request-body.ts
@@ -1,8 +1,0 @@
-export interface AuthRequestBody {
-  responseType: string;
-  clientId: string;
-  redirectUri: string;
-  state: string;
-  scope: string;
-  request: string;
-}
