--- api-tests/src/interfaces/cri-stub-response.ts
@@ -1,3 +1,0 @@
-export interface CriStubResponse {
-  redirectUri: string;
-}
