--- api-tests/src/interfaces/journey-engine-responses.ts
@@ -1,15 +1,0 @@
-export interface JourneyResponse {
-  journey: string;
-}
-
-export interface PageResponse {
-  page: string;
-}
-
-export interface CriResponse {
-  cri: { id: string; redirectUrl: string };
-}
-
-export interface ClientResponse {
-  client: { redirectUrl: string };
-}
