--- api-tests/src/interfaces/process-cri-callback-request.ts
@@ -1,6 +1,0 @@
-export interface ProcessCriCallbackRequest {
-  authorizationCode: string;
-  credentialIssuerId: string;
-  redirectUri: string;
-  state: string;
-}
