--- api-tests/src/interfaces/return-code.ts
@@ -1,3 +1,0 @@
-export interface ReturnCode {
-  code: string;
-}
