--- api-tests/src/interfaces/token-response.ts
@@ -1,6 +1,0 @@
-export interface TokenResponse {
-  access_token: string;
-  scope: string;
-  token_type: string;
-  expires_in: number;
-}
