--- api-tests/src/types/journey-engine-response.ts
@@ -1,12 +1,0 @@
-import {
-  ClientResponse,
-  CriResponse,
-  JourneyResponse,
-  PageResponse,
-} from "../interfaces/journey-engine-responses.js";
-
-export type JourneyEngineResponse =
-  | JourneyResponse
-  | PageResponse
-  | CriResponse
-  | ClientResponse;
